requirements.txt
bcrypt
python-jose
email-validator
fastapi
passlib[bcrypt]
psycopg2-binary
pydantic
pydantic-settings
pydantic_core
python-dotenv
uvicorn[standard]
python-multipart
PyJWT[crypto]
pytest
httpx 
pytest-asyncio
fastapi[standard]
psycopg[binary]
sqlalchemy[asyncio]
testcontainers[postgres]
alembic[sqlalchemy2]
asyncpg

